New Expose
With a whole new focus implemented which differs from previous plan
(which was also awesome). Now the new Research Question and State of
the Art are done.
@arindam-mahanta You can add your part into Methods, I will give it a
final check and sent it to two professors tonight. By the way…Could you
please help me fixing the reference as some of them are no longer in
the expose? ;-) thanks & see you tomorrow