Updated Simulation
So here is the code I showed on Monday to Prof. Hartanto and Prof.
Stuck. I tried to comment it as clear as possible:
By far the agents are able to perform two different driving behaviors
(drunk and normal) on the same map. You just need to push_back a number
of them into a vector. Two class are made for describing these two
behaviors, the classes itself are inheriting from Vehicle class.

Task to be done:
@ianoshan Please make the map scalable. Ideally we just need to assign
the coordinates (e.g. 400, 200) of desired junctions, and then
everything else will work automatically. I think you need to set up
some variables to replace the existing numbers such that we neither
have to manually set up the accessible junction points, nor do we need
to set up the coordinates in order to draw the map in the future.

@arindam-mahanta
Please investigate if we can have more interesting behavior from the
drunk drivers. You can now define all the details of this behavior
under the class DrunkDriver. Hartanto once mentioned on Monday that the
drunk drivers are still sometime overlapping each other. It needs also
to be fixed.

@TimperatorTW
I will be working on the other behavior - normal driver, as now they
are also not implemented perfectly ( also overlapping). And I will try
to introduce maybe a counter or a timer so that we can track the time
spent by the driver travel from A to B.